Harden AM003 assignable collection boundary
